+++ R/calc_1cmt.R
@@ -10,6 +10,7 @@
 #' @return Concentration of drug at requested time (\code{t}) after a single dose, given provided set of parameters and variables.
 #'
 #' @author Justin Wilkins, \email{justin.wilkins@@occams.com}
+#' @author Bill Denney, \email{wdenney@@humanpredictions.com}
 #' @references Bertrand J & Mentre F (2008). Mathematical Expressions of the Pharmacokinetic and Pharmacodynamic Models
 #' implemented in the Monolix software. \url{http://lixoft.com/wp-content/uploads/2016/03/PKPDlibrary.pdf}
 #' @references Rowland M, Tozer TN. Clinical Pharmacokinetics and Pharmacodynamics: Concepts and Applications (4th). Lippincott Williams & Wilkins, Philadelphia, 2010.  
@@ -108,6 +109,7 @@ calc_sd_1cmt_linear_oral_0_lag <- function(t, dose, dur, ...) {
 #' @return Concentration of drug at requested time (\code{t}) after a single dose, given provided set of parameters and variables.
 #'
 #' @author Justin Wilkins, \email{justin.wilkins@@occams.com}
+#' @author Bill Denney, \email{wdenney@@humanpredictions.com}
 #' @references Bertrand J & Mentre F (2008). Mathematical Expressions of the Pharmacokinetic and Pharmacodynamic Models
 #' implemented in the Monolix software. \url{http://lixoft.com/wp-content/uploads/2016/03/PKPDlibrary.pdf}
 #' @references Rowland M, Tozer TN. Clinical Pharmacokinetics and Pharmacodynamics: Concepts and Applications (4th). Lippincott Williams & Wilkins, Philadelphia, 2010.