+++ R/calc_derived.R
@@ -12,7 +12,12 @@
 #' @param Q3 Intercompartmental clearance from central to second peripheral
 #'   compartment (volume per time units, e.g. L/h)
 #' @param ka Absorption rate (inverse time units, e.g. 1/h)
+#' @param dur Duration of zero-order absorption (time units, e.g. h)
 #' @param tlag Absorption lag time (time units, e.g. h)
+#' @param tinf Duration of infusion (time units, e.g. h)
+#' @param dose Dose (amount units, e.g. mg)
+#' @param tau Duration of interdose interval (time units, e.g. h; defaults to 24)
+#' @param step Time increment to use when estimating NCA parameters (time units, e.g. h; defaults to 0.1)
 #' @param type Parameters to return. Default is \code{"all"}.  If not
 #'   \code{"all"}, this may be a vector from the names of the return value list.
 #' @param sigdig Number of significant digits to be returned. Default is
@@ -23,7 +28,8 @@
 #' 
 #' @return Return a list of derived PK parameters for a 1-, 2-, or 3-compartment
 #'   linear model given provided clearances and volumes based on the
-#'   \code{type}.
+#'   \code{type}. If a dose is provided, estimated non-compartmental analysis (NCA) parameters will
+#'   be provided as well, based on simulation of single-dose and (if `tau` is specified) steady-state time courses.
 #' \itemize{ 
 #'   \item \code{Vss}: Volume of distribution at steady state, \eqn{V_{ss}} (volume units, e.g. L); 1-, 2-, and 3-compartment
 #'   \item \code{k10}: First-order elimination rate, \eqn{k_{10}} (inverse time units, e.g. 1/h); 1-, 2-, and 3-compartment
@@ -43,14 +49,25 @@
 #'   \item \code{fracA}: fractional A; 1-, 2-, and 3-compartment
 #'   \item \code{fracB}: fractional B; 2- and 3-compartment
 #'   \item \code{fracC}: fractional C; 3-compartment
+#'   \item \code{AUCinf}: Area under the concentration-time curve to infinity (single dose)
+#'   \item \code{AUCtau}: Area under the concentration-time curve over the dosing interval at steady state
+#'   \item \code{Cmax}: Maximum concentration after a single dose
+#'   \item \code{Cmaxss}: Maximum concentration over the dosing interval at steady state
+#'   \item \code{Tmax}: Time after dose of maximum concentration 
+#'   \item \code{AUCinf_dose_normalized}: Dose-normalized area under the concentration-time curve to infinity (single dose)
+#'   \item \code{AUCtau_dose_normalized}: Dose-normalized area under the concentration-time curve over the dosing interval at steady state
+#'   \item \code{Cmax_dose_normalized}: Dose-normalized maximum concentration after a single dose
+#'   \item \code{Cmaxss_dose_normalized}: Dose-normalized maximum concentration over the dosing interval at steady state
+#'   \item \code{step}: Time increment used when estimating NCA parameters.
 #'  }
 #'
-#' The input parameters with standardized names (\code{V1}, \code{V2},
+#' The input parameters with standardized names (\code{dose}, \code{V1}, \code{V2},
 #' \code{V3}, \code{CL}, \code{Q2}, and \code{Q3}) are also returned in the
-#' list, and if provided, additional PK parameters of `ka` and `lag` are also
+#' list, and if provided, additional PK parameters of `ka`, `tlag`, `tinf` and `dur` are also
 #' returned in the list.  All inputs may be scalars or vectors.
 #'
 #' @author Justin Wilkins, \email{justin.wilkins@@occams.com}
+#' @author Bill Denney, \email{wdenney@@humanpredictions.com}
 #' @references Shafer S. L. \code{CONVERT.XLS}
 #' @references Rowland M, Tozer TN. Clinical Pharmacokinetics and
 #'   Pharmacodynamics: Concepts and Applications (4th). Lippincott Williams &
@@ -58,6 +75,7 @@
 #'
 #' @examples
 #' params <- calc_derived(CL=29.4, V1=23.4, V2=114, V3=4614, Q2=270, Q3=73)
+#' @importFrom PKNCA pk.calc.auc.all
 #' @export
 calc_derived <- function(..., verbose=FALSE) {
   arg_names <- names(list(...))
@@ -83,7 +101,7 @@ calc_derived <- function(..., verbose=FALSE) {
 #' @examples
 #' params <- calc_derived_1cpt(CL=16, V=25)
 #' @export
-calc_derived_1cpt <- function(CL, V=NULL, V1=NULL, ka=NULL, tlag=NULL, type="all", sigdig=5) {
+calc_derived_1cpt <- function(CL, V=NULL, V1=NULL, ka=NULL, dur=NULL, tlag=NULL, tinf=NULL, dose=NULL, tau=NULL, step=0.1, type="all", sigdig=5) {
   if (!xor(is.null(V), is.null(V1))) {
     stop("Exactly one of V or V1 may be provided since they are considered synonyms.")
   } else if (!is.null(V)) {
@@ -95,6 +113,138 @@ calc_derived_1cpt <- function(CL, V=NULL, V1=NULL, ka=NULL, tlag=NULL, type="all
   trueA <- 1/V1
   thalf <- log(2)/k10
   alpha <- k10
+  
+  AUCinf <- NULL
+  AUCtau  <- NULL
+  Cmax   <- NULL
+  Cmaxss <- NULL
+  Tmax   <- NULL
+  AUCinf_dose_normalized <- NULL
+  AUCtau_dose_normalized <- NULL
+  Cmax_dose_normalized   <- NULL
+  Cmaxss_dose_normalized <- NULL
+  
+  guess_tmax <- 0
+  if(!is.null(ka)) guess_tmax <- 1/ka
+  if(!is.null(dur)) guess_tmax <- dur
+  if(!is.null(tinf)) guess_tmax <- tinf
+    
+  tend <- 24
+  while(tend<guess_tmax) {
+    tend <- tend*2
+  }
+  
+  if(!is.null(dose)) {
+    message("dose present. NCA parameters will be estimated.")
+    if(!is.null(tau)) message("tau present. Steady state NCA parameters will be estimated.")
+    AUCinf <- dose/CL
+    AUCinf_dose_normalized <- AUCinf/dose
+    
+    # IV bolus
+    if(is.null(dur) & is.null(ka) & is.null(tlag) & is.null(tinf)) {
+      message("1-compartment bolus detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_1cmt_linear_bolus(t=t, dose=dose, CL=CL, V1=V1)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(is.null(dur) & is.null(ka) & is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_1cmt_linear_bolus(tad = t, dose=dose, tau=tau, CL=CL, V1=V1)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    
+    # IV infusion
+    if(is.null(dur) & is.null(ka) & is.null(tlag) & !is.null(tinf)) {
+      message("1-compartment infusion detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_1cmt_linear_infusion(t=t, dose=dose, CL=CL, V1=V1, tinf = tinf)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(is.null(dur) & is.null(ka) & is.null(tlag) & !is.null(tau) & !is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_1cmt_linear_infusion(tad = t, dose=dose, tau=tau, CL=CL, V1=V1, tinf=tinf)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    # oral first-order
+    if(is.null(dur) & !is.null(ka) & is.null(tlag) & is.null(tinf)) {
+      message("1-compartment first-order oral detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_1cmt_linear_oral_1(t=t, dose=dose, ka=ka, CL=CL, V1=V1)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(is.null(dur) & !is.null(ka) & is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_1cmt_linear_oral_1(tad = t, dose=dose, tau=tau, ka=ka, CL=CL, V1=V1)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    # oral first-order lag
+    if(is.null(dur) & !is.null(ka) & !is.null(tlag) & is.null(tinf)) {
+      message("1-compartment first-order oral with lag time detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_1cmt_linear_oral_1_lag(t=t, dose=dose, ka=ka, CL=CL, V1=V1, tlag=tlag)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(is.null(dur) & !is.null(ka) & !is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_1cmt_linear_oral_1_lag(tad = t, dose=dose, tau=tau, ka=ka, CL=CL, V1=V1, tlag=tlag)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    # oral zero-order
+    if(!is.null(dur) & is.null(ka) & is.null(tlag) & is.null(tinf)) {
+      message("1-compartment zero-order oral detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_1cmt_linear_oral_0(t=t, dose=dose, CL=CL, V1=V1, dur=dur)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(!is.null(dur) & is.null(ka) & is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_1cmt_linear_oral_0(tad = t, dose=dose, tau=tau, dur=dur, CL=CL, V1=V1)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    # oral zero-order lag
+    if(!is.null(dur) & is.null(ka) & !is.null(tlag) & is.null(tinf)) {
+      message("1-compartment zero-order oral with lag time detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_1cmt_linear_oral_0_lag(t=t, dose=dose, CL=CL, V1=V1, tlag=tlag, dur=dur)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(!is.null(dur) & is.null(ka) & !is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_1cmt_linear_oral_0_lag(tad = t, dose=dose, tau=tau, CL=CL, V1=V1, tlag=tlag, dur=dur)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+  }
+  
   out <-
     list(
       k10=signif(CL/V1, sigdig),
@@ -103,11 +253,25 @@ calc_derived_1cpt <- function(CL, V=NULL, V1=NULL, ka=NULL, tlag=NULL, type="all
       alpha=signif(alpha, sigdig),
       trueA=signif(1/V1, sigdig),
       fracA=1,
+      tau=tau,
+      AUCinf=AUCinf,
+      AUCtau=AUCtau,
+      Cmax=Cmax,
+      Cmaxss=Cmaxss,
+      Tmax=Tmax,
+      AUCinf_dose_normalized=AUCinf_dose_normalized,
+      AUCtau_dose_normalized=AUCtau_dose_normalized,
+      Cmax_dose_normalized=Cmax_dose_normalized,
+      Cmaxss_dose_normalized=Cmaxss_dose_normalized,
+      step=step,
       # Include the macro parameters
-      V1=V1,
       CL=CL,
+      V1=V1,
       ka=ka,
-      tlag=tlag
+      tlag=tlag,
+      dur=dur,
+      tinf=tinf,
+      dose=dose
     )
   
   if(type=="all") {
@@ -132,7 +296,7 @@ calc_derived_1cpt <- function(CL, V=NULL, V1=NULL, ka=NULL, tlag=NULL, type="all
 #' @examples
 #' params <- calc_derived_2cpt(CL=16, V1=25, V2=50, Q=0.5)
 #' @export
-calc_derived_2cpt <- function(CL, V1=NULL, V2, Q2=NULL, V=NULL, Q=NULL, ka=NULL, tlag=NULL, type="all", sigdig=5) {
+calc_derived_2cpt <- function(CL, V1=NULL, V2, Q2=NULL, V=NULL, Q=NULL, dur=NULL, tinf=NULL, ka=NULL, tlag=NULL, dose=NULL, tau=NULL, step=0.1, type="all", sigdig=5) {
   if (!xor(is.null(V), is.null(V1))) {
     stop("Exactly one of V or V1 may be provided since they are considered synonyms.")
   } else if (!is.null(V)) {
@@ -160,6 +324,137 @@ calc_derived_2cpt <- function(CL, V1=NULL, V2, Q2=NULL, V=NULL, Q=NULL, ka=NULL,
   c1 <- (k21 - i1)/(i2 - i1)/V1
   c2 <- (k21 - i2)/(i1 - i2)/V1
   
+  AUCinf <- NULL
+  AUCtau  <- NULL
+  Cmax   <- NULL
+  Cmaxss <- NULL
+  Tmax   <- NULL
+  AUCinf_dose_normalized <- NULL
+  AUCtau_dose_normalized <- NULL
+  Cmax_dose_normalized   <- NULL
+  Cmaxss_dose_normalized <- NULL
+  
+  guess_tmax <- 0
+  if(!is.null(ka)) guess_tmax <- 1/ka
+  if(!is.null(dur)) guess_tmax <- dur
+  if(!is.null(tinf)) guess_tmax <- tinf
+  
+  tend <- 24
+  while(tend<guess_tmax) {
+    tend <- tend*2
+  }
+  
+  if(!is.null(dose)) {
+    message("dose present. NCA parameters will be estimated.")
+    if(!is.null(tau)) message("tau present. Steady state NCA parameters will be estimated.")
+    AUCinf <- dose/CL
+    AUCinf_dose_normalized <- AUCinf/dose
+    
+    # IV bolus
+    if(is.null(dur) & is.null(ka) & is.null(tlag) & is.null(tinf)) {
+      message("2-compartment bolus detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_2cmt_linear_bolus(t=t, dose=dose, CL=CL, V1=V1, V2=V2, Q=Q)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(is.null(dur) & is.null(ka) & is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_2cmt_linear_bolus(tad = t, dose=dose, tau=tau, CL=CL, V1=V1, V2=V2, Q=Q)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    
+    # IV infusion
+    if(is.null(dur) & is.null(ka) & is.null(tlag) & !is.null(tinf)) {
+      message("2-compartment infusion detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_2cmt_linear_infusion(t=t, dose=dose, CL=CL, V1=V1, V2=V2, Q=Q, tinf = tinf)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(is.null(dur) & is.null(ka) & is.null(tlag) & !is.null(tau) & !is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_2cmt_linear_infusion(tad = t, dose=dose, tau=tau, CL=CL, V1=V1, V2=V2, Q=Q, tinf=tinf)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    # oral first-order
+    if(is.null(dur) & !is.null(ka) & is.null(tlag) & is.null(tinf)) {
+      message("2-compartment first-order oral detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_2cmt_linear_oral_1(t=t, dose=dose, ka=ka, CL=CL, V1=V1, V2=V2, Q=Q)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(is.null(dur) & !is.null(ka) & is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_2cmt_linear_oral_1(tad = t, dose=dose, tau=tau, ka=ka, CL=CL, V1=V1, V2=V2, Q=Q)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    # oral first-order lag
+    if(is.null(dur) & !is.null(ka) & !is.null(tlag) & is.null(tinf)) {
+      message("2-compartment first-order oral with lag time detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_2cmt_linear_oral_1_lag(t=t, dose=dose, ka=ka, CL=CL, V1=V1, V2=V2, Q=Q, tlag=tlag)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(is.null(dur) & !is.null(ka) & !is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_2cmt_linear_oral_1_lag(tad = t, dose=dose, tau=tau, ka=ka, CL=CL, V1=V1, V2=V2, Q=Q, tlag=tlag)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    # oral zero-order
+    if(!is.null(dur) & is.null(ka) & is.null(tlag) & is.null(tinf)) {
+      message("2-compartment zero-order oral detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_2cmt_linear_oral_0(t=t, dose=dose, CL=CL, V1=V1, V2=V2, Q=Q, dur=dur)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(!is.null(dur) & is.null(ka) & is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_2cmt_linear_oral_0(tad = t, dose=dose, tau=tau, dur=dur, CL=CL, V1=V1, V2=V2, Q=Q)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    # oral zero-order lag
+    if(!is.null(dur) & is.null(ka) & !is.null(tlag) & is.null(tinf)) {
+      message("2-compartment zero-order oral with lag time detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_2cmt_linear_oral_0_lag(t=t, dose=dose, CL=CL, V1=V1, V2=V2, Q=Q, tlag=tlag, dur=dur)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(!is.null(dur) & is.null(ka) & !is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_2cmt_linear_oral_0_lag(tad = t, dose=dose, tau=tau, CL=CL, V1=V1, V2=V2, Q=Q, tlag=tlag, dur=dur)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+  }
+  
   out <-
     list(
       k10=signif(k10, sigdig),
@@ -174,13 +469,27 @@ calc_derived_2cpt <- function(CL, V1=NULL, V2, Q2=NULL, V=NULL, Q=NULL, ka=NULL,
       trueB=signif(c2, sigdig),
       fracA=signif(c1*V1, sigdig),
       fracB=signif(c2*V1, sigdig),
+      tau=tau,
+      AUCinf=AUCinf,
+      AUCtau=AUCtau,
+      Cmax=Cmax,
+      Cmaxss=Cmaxss,
+      Tmax=Tmax,
+      AUCinf_dose_normalized=AUCinf_dose_normalized,
+      AUCtau_dose_normalized=AUCtau_dose_normalized,
+      Cmax_dose_normalized=Cmax_dose_normalized,
+      Cmaxss_dose_normalized=Cmaxss_dose_normalized,
+      step=step,
       # Include the macro parameters
+      CL=CL,
       V1=V1,
       V2=V2,
-      CL=CL,
       Q2=Q2,
       ka=ka,
-      tlag=tlag
+      tlag=tlag,
+      dur=dur,
+      tinf=tinf,
+      dose=dose
     )
   
   if(type=="all") {
@@ -206,7 +515,7 @@ calc_derived_2cpt <- function(CL, V1=NULL, V2, Q2=NULL, V=NULL, Q=NULL, ka=NULL,
 #' @examples
 #' params <- calc_derived_3cpt(CL=29.4, V1=23.4, V2=114, V3=4614, Q2=270, Q3=73)
 #' @export
-calc_derived_3cpt <- function(CL, V1=NULL, V2, V3, Q2=NULL, Q3, V=NULL, Q=NULL, ka=NULL, tlag=NULL, type="all", sigdig=5) {
+calc_derived_3cpt <- function(CL, V1=NULL, V2, V3, Q2=NULL, Q3, V=NULL, Q=NULL, ka=NULL, dur=NULL, tinf=NULL, tlag=NULL, dose=NULL, tau=NULL, step=0.1, type="all", sigdig=5) {
   if (!xor(is.null(V), is.null(V1))) {
     stop("Exactly one of V or V1 may be provided since they are considered synonyms.")
   } else if (!is.null(V)) {
@@ -250,6 +559,138 @@ calc_derived_3cpt <- function(CL, V1=NULL, V2, V3, Q2=NULL, Q3, V=NULL, Q=NULL, 
   c2 <- (k21 - i2) * (k31 - i2) / (i2 - i1) / (i2 - i3) / V1
   c3 <- (k21 - i3) * (k31 - i3) / (i3 - i2) / (i3 - i1) / V1
   
+  AUCinf <- NULL
+  AUCtau  <- NULL
+  Cmax   <- NULL
+  Cmaxss <- NULL
+  Tmax   <- NULL
+  AUCinf_dose_normalized <- NULL
+  AUCtau_dose_normalized <- NULL
+  Cmax_dose_normalized   <- NULL
+  Cmaxss_dose_normalized <- NULL
+  
+  guess_tmax <- 0
+  if(!is.null(ka)) guess_tmax <- 1/ka
+  if(!is.null(dur)) guess_tmax <- dur
+  if(!is.null(tinf)) guess_tmax <- tinf
+  
+  tend <- 24
+  while(tend<guess_tmax) {
+    tend <- tend*2
+  }
+  
+  if(!is.null(dose)) {
+    message("dose present. NCA parameters will be estimated.")
+    if(!is.null(tau)) message("tau present. Steady state NCA parameters will be estimated.")
+    AUCinf <- dose/CL
+    AUCinf_dose_normalized <- AUCinf/dose
+    
+    # IV bolus
+    if(is.null(dur) & is.null(ka) & is.null(tlag) & is.null(tinf)) {
+      message("3-compartment bolus detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_3cmt_linear_bolus(t=t, dose=dose, CL=CL, V1=V1, V2=V2, V3=V3, Q2=Q2, Q3=Q3)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(is.null(dur) & is.null(ka) & is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_3cmt_linear_bolus(tad = t, dose=dose, tau=tau, CL=CL, V1=V1, V2=V2, V3=V3, Q2=Q2, Q3=Q3)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    
+    # IV infusion
+    if(is.null(dur) & is.null(ka) & is.null(tlag) & !is.null(tinf)) {
+      message("3-compartment infusion detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_3cmt_linear_infusion(t=t, dose=dose, CL=CL, V1=V1, V2=V2, V3=V3, Q2=Q2, Q3=Q3, tinf = tinf)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(is.null(dur) & is.null(ka) & is.null(tlag) & !is.null(tau) & !is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_3cmt_linear_infusion(tad = t, dose=dose, tau=tau, CL=CL, V1=V1, V2=V2, V3=V3, Q2=Q2, Q3=Q3, tinf=tinf)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    # oral first-order
+    if(is.null(dur) & !is.null(ka) & is.null(tlag) & is.null(tinf)) {
+      message("3-compartment first-order oral detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_3cmt_linear_oral_1(t=t, dose=dose, ka=ka, CL=CL, V1=V1, V2=V2, V3=V3, Q2=Q2, Q3=Q3)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(is.null(dur) & !is.null(ka) & is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_3cmt_linear_oral_1(tad = t, dose=dose, tau=tau, ka=ka, CL=CL, V1=V1, V2=V2, V3=V3, Q2=Q2, Q3=Q3)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    # oral first-order lag
+    if(is.null(dur) & !is.null(ka) & !is.null(tlag) & is.null(tinf)) {
+      message("3-compartment first-order oral with lag time detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_3cmt_linear_oral_1_lag(t=t, dose=dose, ka=ka, CL=CL, V1=V1, V2=V2, V3=V3, Q2=Q2, Q3=Q3, tlag=tlag)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(is.null(dur) & !is.null(ka) & !is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_3cmt_linear_oral_1_lag(tad = t, dose=dose, tau=tau, ka=ka, CL=CL, V1=V1, V2=V2, V3=V3, Q2=Q2, Q3=Q3, tlag=tlag)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    # oral zero-order
+    if(!is.null(dur) & is.null(ka) & is.null(tlag) & is.null(tinf)) {
+      message("3-compartment zero-order oral detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_3cmt_linear_oral_0(t=t, dose=dose, CL=CL, V1=V1, V2=V2, V3=V3, Q2=Q2, Q3=Q3, dur=dur)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(!is.null(dur) & is.null(ka) & is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_3cmt_linear_oral_0(tad = t, dose=dose, tau=tau, dur=dur, CL=CL, V1=V1, V2=V2, V3=V3, Q2=Q2, Q3=Q3)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+    # oral zero-order lag
+    if(!is.null(dur) & is.null(ka) & !is.null(tlag) & is.null(tinf)) {
+      message("3-compartment zero-order oral with lag time detected.")
+      t <- seq(0, tend, by=step)
+      C <- calc_sd_3cmt_linear_oral_0_lag(t=t, dose=dose, CL=CL, V1=V1, V2=V2, V3=V3, Q2=Q2, Q3=Q3, tlag=tlag, dur=dur)
+      Cmax <- max(C)
+      Tmax <- t[C==Cmax]
+      Cmax_dose_normalized <- Cmax/dose
+    }
+    if(!is.null(dur) & is.null(ka) & !is.null(tlag) & !is.null(tau) & is.null(tinf)) {
+      t <- seq(0, tau, by=step)
+      C <- calc_ss_3cmt_linear_oral_0_lag(tad = t, dose=dose, tau=tau, CL=CL, V1=V1, V2=V2, V3=V3, Q2=Q2, Q3=Q3, tlag=tlag, dur=dur)
+      Cmaxss <- max(C)
+      AUCtau  <- PKNCA::pk.calc.auc.all(conc = C, time = t)
+      AUCtau_dose_normalized <- AUCtau/dose
+      Cmaxss_dose_normalized <- Cmaxss/dose
+    }
+  }
+  
+  
   out <-
     list(
       k10=signif(k10, sigdig),
@@ -275,15 +716,29 @@ calc_derived_3cpt <- function(CL, V1=NULL, V2, V3, Q2=NULL, Q3, V=NULL, Q=NULL, 
       fracA=signif(c1*V1, sigdig),
       fracB=signif(c2*V1, sigdig),
       fracC=signif(c3*V1, sigdig),
+      tau=tau,
+      AUCinf=AUCinf,
+      AUCtau=AUCtau,
+      Cmax=Cmax,
+      Cmaxss=Cmaxss,
+      Tmax=Tmax,
+      AUCinf_dose_normalized=AUCinf_dose_normalized,
+      AUCtau_dose_normalized=AUCtau_dose_normalized,
+      Cmax_dose_normalized=Cmax_dose_normalized,
+      Cmaxss_dose_normalized=Cmaxss_dose_normalized,
+      step=step,
       # Include the macro parameters
+      CL=CL,
       V1=V1,
       V2=V2,
       V3=V3,
-      CL=CL,
       Q2=Q2,
       Q3=Q3,
       ka=ka,
-      tlag=tlag
+      tlag=tlag,
+      dur=dur,
+      tinf=tinf,
+      dose=dose
     )
   if(type=="all") {
     o <- out